Config: Ignore HTTPS errors during requests
We use a self-signed certificate for our domains (since the CA is our
Puppet). However, that results in Chromium HTTPS errors. Given that
Proton cannot communicate with the outside world, it is safe to use the
`ignoreHTTPSErrors` configuration flag since requesting HTML from our
appservers is all the service does.

Change-Id: I0c5bd9324ab0083b75f96d8373bf6d65aea8c975